+++ Cython/Utility/Generator.c
@@ -104,6 +104,7 @@ static PyObject *__Pyx_Generator_Throw(PyObject *gen, PyObject *args);
 static int __Pyx_PyGen_FetchStopIterationValue(PyObject **pvalue) {
     PyObject *et, *ev, *tb;
     PyObject *value = NULL;
+    int result;
 
     __Pyx_ErrFetch(&et, &ev, &tb);
 
@@ -117,6 +118,7 @@ static int __Pyx_PyGen_FetchStopIterationValue(PyObject **pvalue) {
 
     // most common case: plain StopIteration without or with separate argument
     if (likely(et == PyExc_StopIteration)) {
+        int error = 0;
 #if PY_VERSION_HEX >= 0x030300A0
         if (ev && Py_TYPE(ev) == (PyTypeObject*)PyExc_StopIteration) {
             value = ((PyStopIterationObject *)ev)->value;
@@ -128,7 +130,7 @@ static int __Pyx_PyGen_FetchStopIterationValue(PyObject **pvalue) {
             return 0;
         }
 #endif
-        if (!ev || !PyObject_IsInstance(ev, PyExc_StopIteration)) {
+        if (!ev || !(error = PyObject_IsInstance(ev, PyExc_StopIteration))) {
             // PyErr_SetObject() and friends put the value directly into ev
             if (!ev) {
                 Py_INCREF(Py_None);
@@ -139,6 +141,10 @@ static int __Pyx_PyGen_FetchStopIterationValue(PyObject **pvalue) {
             *pvalue = ev;
             return 0;
         }
+        if (unlikely(error == -1)) {
+            // error during isinstance() check
+            return -1;
+        }
     } else if (!PyErr_GivenExceptionMatches(et, PyExc_StopIteration)) {
         __Pyx_ErrRestore(et, ev, tb);
         return -1;
@@ -146,13 +152,19 @@ static int __Pyx_PyGen_FetchStopIterationValue(PyObject **pvalue) {
 
     // otherwise: normalise and check what that gives us
     PyErr_NormalizeException(&et, &ev, &tb);
-    if (unlikely(!PyObject_IsInstance(ev, PyExc_StopIteration))) {
+    result = PyObject_IsInstance(ev, PyExc_StopIteration);
+    if (unlikely(!result)) {
         // looks like normalisation failed - raise the new exception
         __Pyx_ErrRestore(et, ev, tb);
         return -1;
     }
     Py_XDECREF(tb);
     Py_DECREF(et);
+    if (unlikely(result == -1)) {
+        // error during isinstance() check
+        Py_DECREF(ev);
+        return -1;
+    }
 #if PY_VERSION_HEX >= 0x030300A0
     value = ((PyStopIterationObject *)ev)->value;
     Py_INCREF(value);