+++ CHANGES.rst
@@ -81,6 +81,9 @@ Bugs fixed
 
 * Crash when returning values on generator termination.
 
+* In some cases, exceptions raised during internal isinstance() checks were
+  not propagated.
+
 * Runtime reported file paths of source files (e.g for profiling and tracing)
   are now relative to the build root directory instead of the main source file.